Fixed #11 -- Support Django 1.6's ATOMIC_REQUESTS
In Django 1.6 `TransactionMiddleware` was deprecated in favor of
`ATOMIC_REQUESTS`. With this setting activated, all views will be
wrapped in `transaction.atomic`. However the `SessionStore` was
saving sessions within a `transaction.commit_on_success`
wrapper. Mixing old-style and new-style is not supported in all
scenarios, so this resulted in an error.