Fixed bug in updating last_activity
Since the backend only updated the session if it was (marked as) modified, instead of for every page view, the last_activity field doesn't always reflect the actual time of last activity. This change makes the backed always update the session in the database, though possibly, of course, at the cost of an extra query.