Fix geoip on Python 3.x with Django 1.5
When GEOIP_PATH refers directly to a file Django 1.5 on Python 3.x will
not correctly detect the type of the GeoLite database. This is because
Django 1.5 passes a str (not a bytestring) to libgeoip.GeoIP_open().

Referring to the containing folder allows a workaround for this bug -
let Django incorrectly open the database, then reopen it correctly.

Additionally the GeoIP.city() method must be overridden to correct a str
<-> bytes mismatch with the ipv4 test.

This bug was fixed in Django 1.6, and does not affect Django on Python
2.x.

See
- https://code.djangoproject.com/ticket/20384
- https://github.com/django/django/commit/7b00d902