Make templatetags return None instead of 'unknown'
This addresses #67 by dropping the default of returning and instead
returning `None`.  As a result this should probably be seen as a
breaking change, so I'm not sure how thrilled you might be about that.
Additionally, I added support for  to return either the browser or
the os if not both, and added tests for the lot.

I've also updated the session_list template to make use of
`default_if_none` so that it continues to work as before with one
exception: I've replaced the `<i>` tag with `<em>` as the meaning of
`<i>` was [changed in HTML5](https://www.w3.org/TR/html5/text-level-semantics.html#the-i-element).